+++ src/mjlab/asset_zoo/robots/__init__.py
@@ -4,6 +4,15 @@ from mjlab.asset_zoo.robots.i2rt_yam.yam_constants import (
 from mjlab.asset_zoo.robots.i2rt_yam.yam_constants import (
   get_yam_robot_cfg as get_yam_robot_cfg,
 )
+from mjlab.asset_zoo.robots.jljbot.jljbot_constants import (
+  JLJBOT_ACTION_SCALE as JLJBOT_ACTION_SCALE,
+)
+from mjlab.asset_zoo.robots.jljbot.jljbot_constants import (
+  JLJBOT_FOOT_COLLISION_NAMES as JLJBOT_FOOT_COLLISION_NAMES,
+)
+from mjlab.asset_zoo.robots.jljbot.jljbot_constants import (
+  get_jljbot_robot_cfg as get_jljbot_robot_cfg,
+)
 from mjlab.asset_zoo.robots.unitree_g1.g1_constants import (
   G1_ACTION_SCALE as G1_ACTION_SCALE,
 )